upgrade client go version to 0.22.2 and fix int type error

+++ pkg/source/file/job.go
@@ -74,7 +74,7 @@ func JobUid(fileInfo os.FileInfo) string {
 	var buf [64]byte
 	current := strconv.AppendUint(buf[:0], inode, 10)
 	current = append(current, '-')
-	current = strconv.AppendUint(current, device, 10)
+	current = strconv.AppendUint(current, uint64(device), 10)
 	return string(current)
 }
